Parallel beam projector
*kct-pb-krylov implements CVP for parallel ray geometry
*it is the first parallel ray projector in KCT
*new names of data types from CTIOL added

+++ src/CLlin-perfusion.cpp
@@ -94,7 +94,7 @@ public:
                          "pixels count.");
         }
         io::DenSupportedType t = inf.getDataType();
-        if(t != io::DenSupportedType::float_)
+        if(t != io::DenSupportedType::FLOAT32)
         {
             std::string ERR
                 = io::xprintf("This program supports float projections only but the supplied "
